Serqo'meel
You'll probably only hear it here, but the gist to keep the dungeons psy aesthetic while adhering to my slop is that this is a temple dedicated by a contingent of zards that were able to escape persecution during the Gilded to the Ten- of which later were brought under the heel of Psydonic belief- and infused the two together. Basically its an old lore ref.